+++ CHANGELOG.rst
@@ -2,7 +2,6 @@
 Changelog
 =========
 
-
 0.x.x (?)
 ==================
 
@@ -11,6 +10,7 @@ Changelog
 * Added Maximum option for Timeseries
 * Added Number of decimals displays option for Timeseries* Added Bar_Chart_ panel support
 * Extended SqlTarget to support parsing queries from files
+* Fix AlertCondition backwards compatibility (``useNewAlerts`` default to ``False``)
 
 .. _Bar_Chart: basehttps://grafana.com/docs/grafana/latest/panels-visualizations/visualizations/bar-chart/